feat(service-domains): add CRUD ServiceDomainsService

+++ pkg/cachefly/api/services.go
@@ -63,6 +63,10 @@ type EnableAccessLogsRequest struct {
 	LogTarget string `json:"logTarget"`
 }
 
+type EnableOriginLogsRequest struct {
+	LogTarget string `json:"logTarget"`
+}
+
 // Create a new service
 func (s *ServicesService) Create(ctx context.Context, req CreateServiceRequest) (*Service, error) {
 	endpoint := "/services"
@@ -191,6 +195,34 @@ func (s *ServicesService) DeleteAccessLoggingByID(ctx context.Context, id string
 		return nil, fmt.Errorf("id is required")
 	}
 	endpoint := fmt.Sprintf("/services/%s/accessLogs", id)
+
+	var updated Service
+	if err := s.Client.Delete(ctx, endpoint, &updated); err != nil {
+		return nil, err
+	}
+	return &updated, nil
+}
+
+// id is required.
+func (s *ServicesService) EnableOriginLogging(ctx context.Context, id string, req EnableOriginLogsRequest) (*Service, error) {
+	if id == "" {
+		return nil, fmt.Errorf("id is required")
+	}
+	endpoint := fmt.Sprintf("/services/%s/originLogs", id)
+
+	var updated Service
+	if err := s.Client.Put(ctx, endpoint, req, &updated); err != nil {
+		return nil, err
+	}
+	return &updated, nil
+}
+
+// id is required.
+func (s *ServicesService) DeleteOriginLoggingByID(ctx context.Context, id string) (*Service, error) {
+	if id == "" {
+		return nil, fmt.Errorf("id is required")
+	}
+	endpoint := fmt.Sprintf("/services/%s/originLogs", id)
 
 	var updated Service
 	if err := s.Client.Delete(ctx, endpoint, &updated); err != nil {
